Factor out aggregate.
