+++ query-engine/core/src/schema_builder/output_types/mod.rs
@@ -1,3 +1,4 @@
+pub(crate) mod aggregate;
 pub(crate) mod mutation_type;
 pub(crate) mod output_objects;
 pub(crate) mod query_type;
