+++ query-engine/core/src/schema_builder/output_types/query_type.rs
@@ -86,7 +86,7 @@ fn aggregation_field(ctx: &mut BuilderContext, model: &ModelRef) -> OutputField 
     field(
         field_name,
         args,
-        OutputType::object(output_objects::aggregation_object_type(ctx, &model)),
+        OutputType::object(aggregate::aggregation_object_type(ctx, &model)),
         Some(QueryInfo {
             model: Some(Arc::clone(&model)),
             tag: QueryTag::Aggregate,
